# GafferImage::Sampler : Several small optimizations