--- a/include/GafferImage/Sampler.h
+++ b/include/GafferImage/Sampler.h
@@ -107,8 +107,8 @@ class Sampler
 		/// Cached data access
 		/// @param p Any point within the cache that we wish to retrieve the data for.
 		/// @param tileData Is set to the tile's channel data.
-		/// @param tileIndex XY indices that can be used to access the colour value of point 'p' from tileData.
-		inline void cachedData( Imath::V2i p, const float *& tileData, Imath::V2i &tileIndex );
+		/// @param tilePixelIndex XY indices that can be used to access the colour value of point 'p' from tileData.
+		inline void cachedData( Imath::V2i p, const float *& tileData, Imath::V2i &tilePixelIndex );
 
 		const ImagePlug *m_plug;
 		const std::string m_channelName;