These add up to a substantial speedup of of the sample() method.

My simplest test case is this:
```
import GafferImage
import IECore

__children = {}

__children["Resample"] = GafferImage.Resample( "Resample" )
parent.addChild( __children["Resample"] )
__children["Resample"]["filter"].setValue( 'box' )
__children["Resample"]["filterWidth"].setValue( IECore.V2f( 20, 20 ) )
__children["Resample"]["debug"].setValue( 2 )
__children["Constant"] = GafferImage.Constant( "Constant" )
parent.addChild( __children["Constant"] )
__children["Constant"]["format"].setValue( GafferImage.Format( 3000, 3000, 1.000 ) )
__children["Constant"]["color"].setValue( IECore.Color4f( 1, 0.5, 0.1, 1 ) )
__children["Resample"]["in"].setInput( __children["Constant"]["out"] )

del __children
```

The use of the single pass debug mode on Resample with a box filter emphasizes the cost of the sampler, and is a bit like what is going to happen in the warp, where the filter weights can be evaluated in a separable way, but you can't share sampling work with adjacent pixels because the the filter widths are different.

With this test, if I run ```gaffer stats ~/test_crud/warpPerf/slowResample.gfr -image Resample```
I get this with the current code:
```Image generation    22.907s (wall), 361.510s (CPU)```
And this with the new code:
```Image generation    12.154s (wall), 191.540s (CPU)```

In the end, after cleaning it up as much as I can, and removing any optimizations which didn't have a noticeable impact, it seems like a reasonably clean PR.

One of the things I fixed appears to have been an oversight in the original:  clamping the p to the dataWindow should only have been occuring when bounding mode is clamp, but it was happening for all pixels unless they were outside the image.  The thing I'm most expecting you might take issue with was I added a new value of -1 for m_boundingMode - this isn't a publicly visible boundingMode, it just flags that we don't need to think at all about bounding mode, because the sample window is a subset of the data window, so the sampler isn't responsible for any boundary pixels - would you prefer this as a separate bool?

I think most of the changes are moderately self-explanatory, and the tests pass, but this is some fairly core code, so it's definitely worth you taking a look through and making sure it makes sense and I haven't missed anything.