@johnhaddon 2017-03-07
Thanks Daniel - good stuff! I've added a binding a test for `tileIndex()` to put my mind at ease about the results for negative positions, and merged.

If you feel like tweaking things any more as you go on, I've always thought it's awkward that `cachedData()` returns void and then uses two reference parameters as outputs - perhaps the `float *` could just be returned, leaving just one output parameter?